improve pkg extraction logic
When extracting an attr path from a store path in process:

- strp until first `-`, included
- leave version number
- strip `.drv` if present

This is so that we can do more precise detection of e.g. Python
packages.

Those end up in the store with a path like /nix/store/foobar-python3.12-babel-1.2.3.drv,
so simply splitting on `-` and taking the second fragment like we did before doesn't work
(since it would only take python3.12).

Leaving the version number in is just a convenience (one less thing to
worry about). It does require switching from `.pname` to `.name` in the
jq query though.